돈나가기 추가

+++ model/jdbc/Transfer/Transfer_logDAO.java
@@ -40,7 +40,7 @@ public class Transfer_logDAO {
 				dto.setTarget(rs.getString("target"));
 				dto.setTo_account_number("to_account_number");
 				dto.setReceived(rs.getString("received"));
-				dto.setSum(rs.getInt("sum"));
+				dto.setSum(rs.getLong("sum"));
 				dto.setFee(rs.getInt("fee"));
 				dto.setCms(rs.getString("cms"));
 				dto.setMemo(rs.getString("memo"));
@@ -60,7 +60,7 @@ public class Transfer_logDAO {
 				dto.setTarget(rs.getString("target"));
 				dto.setTo_account_number("to_account_number");
 				dto.setReceived(rs.getString("received"));
-				dto.setSum(rs.getInt("sum"));
+				dto.setSum(rs.getLong("sum"));
 				dto.setFee(rs.getInt("fee"));
 				dto.setCms(rs.getString("cms"));
 				dto.setMemo(rs.getString("memo"));
@@ -122,7 +122,7 @@ public class Transfer_logDAO {
 			pstmt.setString(3, dto.getTarget());
 			pstmt.setString(4, dto.getTo_account_number());
 			pstmt.setString(5, dto.getReceived());
-			pstmt.setInt(6, dto.getSum());
+			pstmt.setLong(6, dto.getSum());
 			pstmt.setInt(7, dto.getFee());
 			pstmt.setString(8, dto.getCms());
 			pstmt.setString(9, dto.getMemo());
